+++ packages/storybook-utils/src/preview.spec.ts
@@ -6,11 +6,12 @@ import { replaceAll, sourceCodeTransformer } from "./preview";
 import * as sourceCodeGenerator from "./source-code-generator";
 
 describe("preview.ts", () => {
-  test("should transform source code and add icon imports", () => {
+  test("should transform source code and add icon/onyx imports", () => {
     // ARRANGE
     const generatorSpy = vi.spyOn(sourceCodeGenerator, "generateSourceCode")
       .mockReturnValue(`<template>
 <OnyxTest icon='${placeholder}' test='${bellRing}' :obj="{foo:'${replaceAll(calendar, '"', "\\'")}'}" />
+<OnyxOtherComponent />
 </template>`);
 
     // ACT
@@ -19,6 +20,7 @@ describe("preview.ts", () => {
     // ASSERT
     expect(generatorSpy).toHaveBeenCalledOnce();
     expect(sourceCode).toBe(`<script lang="ts" setup>
+import { OnyxOtherComponent, OnyxTest } from "sit-onyx";
 import bellRing from "@sit-onyx/icons/bell-ring.svg?raw";
 import calendar from "@sit-onyx/icons/calendar.svg?raw";
 import placeholder from "@sit-onyx/icons/placeholder.svg?raw";
@@ -26,6 +28,7 @@ import placeholder from "@sit-onyx/icons/placeholder.svg?raw";
 
 <template>
 <OnyxTest :icon="placeholder" :test="bellRing" :obj="{foo:calendar}" />
+<OnyxOtherComponent />
 </template>`);
   });
 });